write calculation EuTaxonomyShare calculation rule
