+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/services/datapoints/DataPointConversion.kt
@@ -5,6 +5,7 @@ import org.dataland.datalandbackend.frameworks.eutaxonomynonfinancials.custom.Eu
 import org.dataland.datalandbackend.model.datapoints.ExtendedDataPoint
 import org.dataland.datalandbackend.model.datapoints.UploadedDataPoint
 import org.dataland.datalandbackend.model.datapoints.extended.ExtendedCurrencyDataPoint
+import org.dataland.datalandbackend.model.datapoints.extended.ExtendedDecimalDataPoint
 import org.dataland.datalandbackendutils.model.DataPointType
 import org.dataland.specificationservice.openApiClient.model.DataPointTypeSpecification
 import org.dataland.specificationservice.openApiClient.model.FrameworkSpecification
@@ -111,6 +112,22 @@ enum class DataPointConversion(
             specs: Map<DataPointType, DataPointTypeSpecification>,
             sourceFrameworksByType: Map<DataPointType, List<FrameworkSpecification>>,
         ): UploadedDataPoint = convertEuTaxonomyActivityMerge(inputs, targetType, specs, sourceFrameworksByType)
+    },
+
+    EU_TAXONOMY_SHARE("EuTaxonomyShare") {
+        override fun createComment(
+            inputs: Collection<UploadedDataPoint>,
+            specs: Map<DataPointType, DataPointTypeSpecification>,
+            dataPoints: Collection<ExtendedDataPointInterface<*>>,
+            sourceFrameworksByType: Map<DataPointType, List<FrameworkSpecification>>,
+        ): String = createCommentEuTaxonomyShare(inputs, specs, dataPoints, sourceFrameworksByType)
+
+        override fun convert(
+            inputs: Collection<UploadedDataPoint>,
+            targetType: DataPointType,
+            specs: Map<DataPointType, DataPointTypeSpecification>,
+            sourceFrameworksByType: Map<DataPointType, List<FrameworkSpecification>>,
+        ): UploadedDataPoint = convertEuTaxonomyShare(inputs, targetType, specs, sourceFrameworksByType)
     }, ;
 
     /**
@@ -337,6 +354,38 @@ enum class DataPointConversion(
                         listOfNotNull(activityLists.alignedActivities, activityLists.nonAlignedActivities),
                         sourceFrameworksByType,
                     ),
+                dataSource = mergeDataSources(sources.mapNotNull { it?.let(::getDataSource) }),
+            )
+        return createUploadedDataPoint(
+            inputs = inputs,
+            targetType = targetType,
+            calculatedDataPoint = calculatedDataPoint,
+        )
+    }
+
+    /**
+     * Derives the share of the eligible or aligned activities of a single activity group from the eligible-or-aligned
+     * activity list of the EU taxonomy 2026/73 framework.
+     *
+     * @param inputs the single source data point holding the eligible-or-aligned activity list
+     * @param targetType the data point type assigned to the resulting data point
+     * @param specs the data point type specifications used to resolve the input's role
+     * @param sourceFrameworksByType framework specifications associated with each source data point type
+     * @return the derived share as an [ExtendedDecimalDataPoint]
+     */
+    protected fun convertEuTaxonomyShare(
+        inputs: Collection<UploadedDataPoint>,
+        targetType: DataPointType,
+        specs: Map<DataPointType, DataPointTypeSpecification>,
+        sourceFrameworksByType: Map<DataPointType, List<FrameworkSpecification>>,
+    ): UploadedDataPoint {
+        val operands = extractEuTaxonomyActivitiesOperands(inputs, specs)
+        val sources = operands.sources
+        val calculatedDataPoint =
+            ExtendedDecimalDataPoint(
+                value = operands.calculateShare(resolveEuTaxonomyShareRule(targetType)),
+                quality = mergeQuality(sources.map { it?.quality }),
+                comment = createComment(inputs, specs, sources.filterNotNull(), sourceFrameworksByType),
                 dataSource = mergeDataSources(sources.mapNotNull { it?.let(::getDataSource) }),
             )
         return createUploadedDataPoint(
